fix: allow type name to be a new atom
When using custom types via an Ash.Type storage_type impl, and returning
something like `nvarchar(100)`, and we then end up trying to create an
atom for `nvarchar` (without the size). That atom may not exist, and we
should just be OK with that because we're generating migrations and the
types here are bounded by the content of our code anyhow.